refactor(core): extract chat composer service with port-based DI
Move send/abort/slash-command logic from ChatInput.tsx and App.tsx
into createChatComposer factory in @clawwork/core. Split ChatInput
(1655 LOC) into 9 focused modules with extracted hooks. Desktop
renderer no longer calls window.clawwork.sendMessage or
window.clawwork.persistMessage directly — all user message writes
flow through the composer, preserving the single-writer invariant.
