+++ packages/core/src/index.ts
@@ -82,3 +82,6 @@ export { autoTitleIfNeeded } from './services/auto-title.js';
 
 export { createGatewayDispatcher } from './services/gateway-dispatcher.js';
 export type { GatewayDispatcherDeps } from './services/gateway-dispatcher.js';
+
+export { createChatComposer } from './services/chat-composer.js';
+export type { ChatComposerDeps, ChatComposer, SendOptions } from './services/chat-composer.js';
