+++ packages/desktop/src/renderer/components/ChatInput/types.ts
@@ -1,0 +1,4 @@
+export interface PendingImage {
+  file: File;
+  previewUrl: string;
+}
